+++ projects/pioneer-charts/src/lib/bar-chart/bar-vertical-chart/bar-vertical-chart.builder.ts
@@ -12,7 +12,7 @@ import 'd3-transition';
  */
 import { PcacBarVerticalChartConfig } from './bar-vertical-chart.model';
 import { PcacChart } from '../../core/chart';
-import { PcacData, PcacFormatEnum } from '../../core/chart.model';
+import { PcacData } from '../../core/chart.model';
 import { stackStarts } from '../../core/stack';
 
 import { Subject } from 'rxjs';
@@ -63,9 +63,10 @@ export class BarVerticalChartBuilder extends PcacChart {
   private buildScales(config: PcacBarVerticalChartConfig) {
     // `this.height`, not `config.height`: with `heightFull` on they differ (see
     // PcacChart.resolveHeight), and the axis, grid and bar baselines all use the resolved one.
+    // Bars grow from 0, so only `domainMax` is read; `yAxis.domainMin` is ignored.
     this.yScale = scaleLinear()
       .rangeRound([0, this.height])
-      .domain([config.domainMax, 0]);
+      .domain([this.yAxis.domainMax as number, 0]);
 
     this.xScaleStacked = scaleBand()
       .domain(config.data.map((d) => d.key as string))
@@ -80,7 +81,7 @@ export class BarVerticalChartBuilder extends PcacChart {
 
   private drawChart(chartElm: ElementRef, config: PcacBarVerticalChartConfig): void {
     this.buildContainer(chartElm);
-    this.axisBuilder.drawAxis(this.axisBuilderConfig(this.xScaleStacked, this.yScale, PcacFormatEnum.None, config.tickFormat || PcacFormatEnum.None));
+    this.axisBuilder.drawAxis(this.axisBuilderConfig(this.xScaleStacked, this.yScale));
     this.drawGrids(this.xScaleStacked, this.yScale);
     this.addGroups(config);
     this.axisBuilder.raiseAxes(this.svg);
@@ -175,7 +176,7 @@ export class BarVerticalChartBuilder extends PcacChart {
       .on('mousemove', function (this: SVGRectElement, event: MouseEvent, d: PcacData) {
         const groupIndex = Number((this.parentNode as Element).getAttribute('data-group-id'));
         const index = Number(this.getAttribute('data-group-bar-id'));
-        self.showTooltip(event, d, { index, parent: config.data[groupIndex], parentIndex: groupIndex, valueFormat: config.tickFormat || PcacFormatEnum.None });
+        self.showTooltip(event, d, { index, parent: config.data[groupIndex], parentIndex: groupIndex, valueFormat: self.yAxis.format });
       })
       .on('mouseout', function (this: any) {
         self.hideTooltip();
@@ -215,7 +216,7 @@ export class BarVerticalChartBuilder extends PcacChart {
         return i;
       })
       .on('mousemove', (event: MouseEvent) => {
-        this.showTooltip(event, config.thresholds[0], { index: 0, isThreshold: true, valueFormat: config.tickFormat || PcacFormatEnum.None });
+        this.showTooltip(event, config.thresholds[0], { index: 0, isThreshold: true, valueFormat: this.yAxis.format });
       })
       .transition()
       .duration(this.transitionService.getTransitionDuration())
@@ -239,7 +240,7 @@ export class BarVerticalChartBuilder extends PcacChart {
         self.showTooltip(
           event,
           config.isStacked ? threshold.data[0] : threshold,
-          { index, parent: config.data[index], parentIndex: index, isThreshold: true, valueFormat: config.tickFormat || PcacFormatEnum.None }
+          { index, parent: config.data[index], parentIndex: index, isThreshold: true, valueFormat: self.yAxis.format }
         );
       })
       .transition()
@@ -263,7 +264,7 @@ export class BarVerticalChartBuilder extends PcacChart {
         self.showTooltip(
           event,
           config.thresholds[groupIndex].data[index],
-          { index, parent: config.data[groupIndex], parentIndex: groupIndex, isThreshold: true, valueFormat: config.tickFormat || PcacFormatEnum.None }
+          { index, parent: config.data[groupIndex], parentIndex: groupIndex, isThreshold: true, valueFormat: self.yAxis.format }
         );
       })
       .attr('width', this.xScaleGrouped.bandwidth())