+++ projects/pioneer-charts/src/lib/bar-chart/bar-vertical-chart/bar-vertical-chart.model.ts
@@ -1,10 +1,8 @@
-import { PcacAxisChartConfig, PcacData, PcacFormatEnum } from '../../core/chart.model';
+import { PcacAxisChartConfig, PcacData } from '../../core/chart.model';
 
 export class PcacBarVerticalChartConfig extends PcacAxisChartConfig {
-  domainMax: number = 100;
   isStacked: boolean = false;
   thresholds: PcacData[] = [];
-  tickFormat: PcacFormatEnum = PcacFormatEnum.None
   spreadColorsPerGroup: boolean = false
   colorOverride: PcacBarVerticalChartColorOverrideConfig = new PcacBarVerticalChartColorOverrideConfig()
 }